backend/booking-service/requirements.txt
fastapi[all]
mysql-connector-python
pydantic
